+++ CHANGELOG.md
@@ -373,6 +373,7 @@
 * ActionSpecialSplitter now catches exceptions in called action specials by default
 * ActorController now catches exceptions thrown by OnDeathSpecial
 * GetParamsForModule now checks if there's actually a delimiter after the module name
+* GetActiveCamera getting player camera is now delegated to PlayerController via IControlPlayerCamera interface
 * Added conditional/microscript resolvers for PlayerFlags and SessionFlags (conditional only) [Experimental]
 * Added ScriptStringSubber that calls scripts and substitutes the result [Experimental]
 * DelayScaled can now be used from threads other than the main thread [Untested]